Addressing comments and moving stuff to xml
