+++ BeeTracker/app/src/main/java/com/blueprint/foe/beetracker/IdentifyFragment.java
@@ -14,6 +14,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.blueprint.foe.beetracker.Listeners.OnBeePartSelectedListener;
 import com.blueprint.foe.beetracker.Model.BeePart;
 import com.blueprint.foe.beetracker.Model.PartsPickerAdapter;
 import com.blueprint.foe.beetracker.Model.Submission;
