+++ BeeTracker/app/src/main/java/com/blueprint/foe/beetracker/Listeners/BeeAlertDialogListener.java
@@ -1,0 +1,8 @@
+package com.blueprint.foe.beetracker.Listeners;
+
+/**
+ * The listener for dialog boxes
+ */
+public interface BeeAlertDialogListener {
+    void onDialogFinishClick();
+}
