+++ BeeTracker/app/src/main/java/com/blueprint/foe/beetracker/Listeners/OnBeePartSelectedListener.java
@@ -1,4 +1,4 @@
-package com.blueprint.foe.beetracker;
+package com.blueprint.foe.beetracker.Listeners;
 
 /**
  * This interface lets the BeePickerAdapter tell the IdentifyFragment that an item has been
