+++ BeeTracker/app/src/main/res/mipmap-mdpi/icon_done_checkmark_popup.png
