+++ BeeTracker/app/src/main/res/values/strings.xml
@@ -15,7 +15,7 @@
     <string name="face">FACE</string>
     <string name="thorax">THORAX</string>
     <string name="abdomen">ABDOMEN</string>
-    <string name="submit_dialog_heading">Buzz Buzz!</string>
+    <string name="submit_dialog_heading">Buzz buzz!</string>
     <string name="submit_dialog_paragraph">(That\'s thank you in bee!)</string>
     <string name="home">Home</string>
     <string name="learn">Learn</string>
